disable the sh4 cache and delete itoa
itoa is a function I wrote for debugging purposes and left in by accident.

If I delete itoa without disabling the cache, then the test will fail on real
hardware.  I'm not sure what's going on, but somehow that function's existance
was interfering with the cache in a good way.  Maybe it was causing the packet
to end up in a cache line that's subject to some sort of contention?

The most confusing part is I thought the 0xa0000000 address prefix was supposed
to bypass the cache but that doesn't seem to be the case.

I'd like to go back and fix this the right way (without disabling cache), but
for now I just need things to work.